feat: 기획 변경에 따른 라우팅 및 레이아웃 수정
