+++ src/features/posts/components/Post/index.ts
@@ -1,0 +1,1 @@
+export * from "./Post";
