+++ src/features/posts/components/Posts/Posts.tsx
@@ -1,0 +1,18 @@
+import { VStack } from "@chakra-ui/react";
+import React from "react";
+
+import { Post } from "../Post";
+
+export const Posts = () => {
+  return (
+    <>
+      <VStack spacing={4}>
+        <Post />
+        <Post />
+        <Post />
+        <Post />
+        <Post />
+      </VStack>
+    </>
+  );
+};
