+++ src/features/posts/components/Posts/index.ts
@@ -1,0 +1,1 @@
+export * from "./Posts";
