+++ src/features/posts/components/PostsSearchForm/index.ts
@@ -1,0 +1,1 @@
+export * from "./PostsSearchForm";
