+++ src/features/posts/components/SortBySelect/index.ts
@@ -1,0 +1,1 @@
+export * from "./SortBySelect";
