+++ src/features/posts/components/index.ts
@@ -1,0 +1,4 @@
+export * from "./Post";
+export * from "./Posts";
+export * from "./PostsSearchForm";
+export * from "./SortBySelect";
