--- src/pages/connection/ConnectionPage.tsx
@@ -1,9 +1,0 @@
-import { PageLayout } from "@/components/Layout";
-
-export const ConnectionPage = () => {
-  return (
-    <PageLayout title="동행" header={<div>동행</div>}>
-      Connection
-    </PageLayout>
-  );
-};
