+++ src/pages/index.ts
@@ -1,4 +1,3 @@
-export * from "./connection";
 export * from "./posts";
 export * from "./profile";
-export * from "./schedule";
+export * from "./schedules";
