+++ src/pages/posts/PostsPage.tsx
@@ -1,21 +1,26 @@
-import { PostsPageHeader, ConnectionPosts, SchedulePosts } from "./components";
-import { usePostsPage } from "./hooks/usePostsPage";
+import { Box, Flex, HStack, Heading } from "@chakra-ui/react";
 
 import { PageLayout } from "@/components/Layout";
-
-const ContentComponents = [SchedulePosts, ConnectionPosts];
+import {
+  Posts,
+  PostsSearchForm,
+  SortBySelect,
+} from "@/features/posts/components";
 
 export const PostsPage = () => {
-  const { contentIndex, handleChange } = usePostsPage();
-
-  const ContentComponent = ContentComponents[contentIndex];
-
   return (
-    <PageLayout
-      title="게시글"
-      header={<PostsPageHeader index={contentIndex} onChange={handleChange} />}
-    >
-      <ContentComponent />
+    <PageLayout title="플랜찾기" header={<Heading size="lg">플랜찾기</Heading>}>
+      <Box my={4}>
+        <PostsSearchForm />
+      </Box>
+      <Flex justify="flex-end" my={4}>
+        <HStack flexShrink={0}>
+          <SortBySelect variant="unstyled" />
+        </HStack>
+      </Flex>
+      <Box my={4}>
+        <Posts />
+      </Box>
     </PageLayout>
   );
 };
