--- src/pages/posts/components/ConnectionPosts/ConnectionPosts.tsx
@@ -1,3 +1,0 @@
-export const ConnectionPosts = () => {
-  return <div>connection</div>;
-};
