--- src/pages/posts/components/ConnectionPosts/index.ts
@@ -1,1 +1,0 @@
-export * from "./ConnectionPosts";
