--- src/pages/posts/components/PostsPageHeader/index.ts
@@ -1,1 +1,0 @@
-export * from "./PostsPageHeader";
