--- src/pages/posts/components/SchedulePosts/index.ts
@@ -1,1 +1,0 @@
-export * from "./SchedulePosts";
