--- src/pages/posts/components/SchedulePostsSearchForm/index.ts
@@ -1,1 +1,0 @@
-export * from "./SchedulePostsSearchForm";
