--- src/pages/posts/components/index.ts
@@ -1,3 +1,0 @@
-export * from "./ConnectionPosts";
-export * from "./PostsPageHeader";
-export * from "./SchedulePosts";
