--- src/pages/schedule/SchedulePage.tsx
@@ -1,9 +1,0 @@
-import { PageLayout } from "@/components/Layout";
-
-export const SchedulePage = () => {
-  return (
-    <PageLayout title="일정 공유" header={<div>일정 공유</div>}>
-      Schedule
-    </PageLayout>
-  );
-};
