--- src/pages/schedule/index.ts
@@ -1,1 +1,0 @@
-export * from "./SchedulePage";
