+++ src/pages/schedules/index.ts
@@ -1,0 +1,1 @@
+export * from "./SchedulesPage";
